adjustable scope shaking
add adjustable scope shaking.
Add some shoot feed back and effect

+++ Cyberpunk_GameJam/Assets/Script/Target_Info.cs
@@ -17,7 +17,7 @@ public class Target_Info : MonoBehaviour
 
     public BodyParts bodyPart;
 
-    public float score;
+    public int score;
     private int count;
     // Start is called before the first frame update
     void Start()
